vagrant: create config file
With the lack of a development environment, the code may function
differently on different operating systems.

Create a `Vagrantfile`, which sets up a Debian environment, loaded with
Java 11, Maven, and a PostgreSQL database.

Create the schema `tempo`, owned by user `vagrant. This may change in
CI, when deploying to production, but will do for now.