tempo-service: project structure and test code
With a development environment ready, structure the service code as
follows:
- `tempo-parent`
- `tempo-core`
- `tempo-db`

Make the `tempo-parent` module contain depency and plugin management,
for all children to inherit. The most important dependencies are:
- Flyway
- Jetty
- JOOQ
- PostgreSQL

Flyway migrations should go in the `tempo-db` module, under:
- `src/resources/migrations/core/postgres`

It is crucial that the migration filenames follow the naming convention,
`Vxxx__some_name` - replacing `xxx` with a zero-left-padded digit,
followed by TWO underscores and a snake_case name.

JOOQ generates code for the database schema and allows executing queries
by object orientation. This makes it more difficult to make errors.

Create a test table (to be removed) with a test value in the `tempo`
schema.

Create a simple RESTful service, with one GET and one POST operation.
Run a select query and on `core.test` and use it in the above GET
request response.

API requests should be asynchronous, to prevent blocking requests from
other simlutaneous requests.

This code should work well as a foundation for the actual service code.
After modelling the production database and API structure, it should be
easy to add any new API services/operations.